1.01a

+++ runner-extension/fakechange_patcher.sh
@@ -10,15 +10,22 @@ if Patcher/FakeChangePatcher DELTARUNE; then
 	if ! [ -d lib ]; then
 		mkdir lib
 	fi
-	
-	gcc -c -fpic ../FakeChange/RunnerPreload.c
-	gcc -shared -o lib/libfcpreload.so RunnerPreload.o
-	rm RunnerPreload.o
 
-	gcc -c -fpic ../FakeChange/FakeChange.c
-	gcc -shared -o lib/libfakechange.so FakeChange.o
-	rm FakeChange.o
+	if command -v gcc; then
+		echo "Found gcc, compiling libraries..."
+		gcc -c -fpic ../FakeChange/RunnerPreload.c
+		gcc -shared -o lib/libfcpreload.so RunnerPreload.o
+		rm RunnerPreload.o
+
+		gcc -c -fpic ../FakeChange/FakeChange.c
+		gcc -shared -o lib/libfakechange.so FakeChange.o
+		rm FakeChange.o
+	else
+		echo "Copying libraries..."
+		cp ../FakeChange/*.so lib
+	fi
 	cd ..
+	echo "Done"
 	
 else
 	echo "Failed to run patcher"
